requirements.txt
gspread
google-auth
beautifulsoup4
python-dotenv
fastapi
pydanti